requirements.txt
PyQt5==5.11.3
cython==0.29.1
sphinx==1.8.2
matplotlib==3.0.2
nose==1.3.7
numpy==1.15.4
pyserial==3.4
scipy==1.1.0
sphinx-autobuild==0.6.0
virtualenvwrapper
scikit-image==0.14.1
pyzmq==17.1.2
openpyxl==2.5.11
configparser==3.5.0
opencv-python==3.4.0.12
